add delete roles api
