--- infrastructure/src/sd-users/roles-delete.ts
@@ -1,149 +1,0 @@
-import {PythonFunction, PythonFunctionProps} from '@aws-cdk/aws-lambda-python-alpha';
-import {Aws, Duration} from 'aws-cdk-lib';
-import {
-    JsonSchemaType,
-    JsonSchemaVersion,
-    LambdaIntegration,
-    Model,
-    RequestValidator,
-    Resource,
-} from 'aws-cdk-lib/aws-apigateway';
-import {MethodOptions} from 'aws-cdk-lib/aws-apigateway/lib/method';
-import {Table} from 'aws-cdk-lib/aws-dynamodb';
-import {Effect, PolicyStatement, Role, ServicePrincipal} from 'aws-cdk-lib/aws-iam';
-import {Architecture, LayerVersion, Runtime} from 'aws-cdk-lib/aws-lambda';
-import {Construct} from 'constructs';
-
-export interface DeleteRolesApiProps {
-    router: Resource;
-    httpMethod: string;
-    multiUserTable: Table;
-    srcRoot: string;
-    commonLayer: LayerVersion;
-}
-
-export class DeleteRolesApi {
-    private readonly src;
-    private readonly router: Resource;
-    private readonly httpMethod: string;
-    private readonly scope: Construct;
-    private readonly multiUserTable: Table;
-    private readonly layer: LayerVersion;
-    private readonly baseId: string;
-
-    constructor(scope: Construct, id: string, props: DeleteRolesApiProps) {
-        this.scope = scope;
-        this.baseId = id;
-        this.router = props.router;
-        this.httpMethod = props.httpMethod;
-        this.multiUserTable = props.multiUserTable;
-        this.src = props.srcRoot;
-        this.layer = props.commonLayer;
-
-        this.deleteEndpointsApi();
-    }
-
-    private iamRole(): Role {
-
-        const newRole = new Role(this.scope, `${this.baseId}-role`, {
-            assumedBy: new ServicePrincipal('lambda.amazonaws.com'),
-        });
-
-        newRole.addToPolicy(new PolicyStatement({
-            actions: [
-                'dynamodb:Query',
-                'dynamodb:GetItem',
-                'dynamodb:PutItem',
-                'dynamodb:DeleteItem',
-                'dynamodb:Describe*',
-                'dynamodb:List*',
-                'dynamodb:Scan',
-            ],
-            resources: [
-                this.multiUserTable.tableArn,
-            ],
-        }));
-
-        newRole.addToPolicy(new PolicyStatement({
-            effect: Effect.ALLOW,
-            actions: [
-                'logs:CreateLogGroup',
-                'logs:CreateLogStream',
-                'logs:PutLogEvents',
-            ],
-            resources: [`arn:${Aws.PARTITION}:logs:${Aws.REGION}:${Aws.ACCOUNT_ID}:log-group:*:*`],
-        }));
-
-        return newRole;
-    }
-
-    private deleteEndpointsApi() {
-
-        const lambdaFunction = new PythonFunction(this.scope, `${this.baseId}-lambda`, <PythonFunctionProps>{
-            entry: `${this.src}/roles`,
-            architecture: Architecture.X86_64,
-            runtime: Runtime.PYTHON_3_9,
-            index: 'delete_roles.py',
-            handler: 'handler',
-            timeout: Duration.seconds(900),
-            role: this.iamRole(),
-            memorySize: 1024,
-            environment: {
-                MULTI_USER_TABLE: this.multiUserTable.tableName,
-            },
-            layers: [this.layer],
-        });
-
-        const model = new Model(this.scope, `${this.baseId}-model`, {
-            restApi: this.router.api,
-            modelName: 'DeleteRolesModel',
-            description: 'Delete Roles Model',
-            schema: {
-                schema: JsonSchemaVersion.DRAFT4,
-                title: 'deleteRolesSchema',
-                type: JsonSchemaType.OBJECT,
-                properties: {
-                    roles: {
-                        type: JsonSchemaType.ARRAY,
-                        items: {
-                            type: JsonSchemaType.STRING,
-                        },
-                        minItems: 1,
-                        maxItems: 10,
-                    },
-                },
-                required: [
-                    'roles',
-                ],
-            },
-            contentType: 'application/json',
-        });
-
-        const deleteRolesIntegration = new LambdaIntegration(
-            lambdaFunction,
-            {
-                proxy: false,
-                integrationResponses: [{statusCode: '200'}],
-            },
-        );
-
-        const requestValidator = new RequestValidator(this.scope, `${this.baseId}-validator`, {
-            restApi: this.router.api,
-            requestValidatorName: 'DeleteRolesRequestValidator',
-            validateRequestBody: true,
-        });
-
-        this.router.addMethod(this.httpMethod, deleteRolesIntegration, <MethodOptions>{
-            apiKeyRequired: true,
-            requestValidator,
-            requestModels: {
-                'application/json': model,
-            },
-            methodResponses: [
-                {statusCode: '200'},
-                {statusCode: '500'},
-            ],
-        });
-
-    }
-}
